gitvote: Do not exclude team maintainers from voting
Some Steering members are also GitHub organization owners, which
will automatically designate us as maintainers of any team we hold
membership in (like `gitvote-steering`).

This config change allows the subset of Steering members who are
also GitHub org owners to have binding votes.

Opting to explicitly set this config option to `false`, instead of
assuming this will always be the default, so we don't get surprised
in the future.

Signed-off-by: Stephen Augustus <foo@auggie.dev>