gitvote: Update pass threshold to a simple majority (#330)
The pass threshold for a vote was set to 67 to account for the fact
that Leslie was not a member of the GitHub organization.

Now that Leslie has joined the org, this updates the gitvote pass
threshold to align with the simple majority (51%) or four (4) passing
electronic votes required by our charter, at this point in time.

Signed-off-by: Stephen Augustus <foo@auggie.dev>